refactor(cli): refactor cursor blink useEffect into the useEditBuffer custom hook as per React best practices
